Fix ToolShortcuts on latest game version
In older versions the mod could not be loaded, because an interface of the game changed.
This has been updated, now the mod loads again.

Additional, the BepInEx version has been changed to be the same that the official Timberborn-BepInEx-Mod has. Removed the wildcard '5*', because the dependencies in the manifest file had a constant version too.

Finally bumped the version to 3.2 as there was a change and the change should be released.